+++ includes/class-polygon-tweaks.php
@@ -92,8 +92,9 @@ class Polygon_Tweaks {
 		add_action( 'after_setup_theme', array( $admin, 'add_image_sizes' ) );
 		add_action( 'intermediate_image_sizes', array( $admin, 'remove_image_sizes' ), 999 );
 		add_filter( 'big_image_size_threshold', array( $admin, 'change_big_image_threshold' ), 999 );
-		add_filter( 'jpeg_quality', array( $admin, 'change_image_quality' ) );
+		add_filter( 'max_srcset_image_width', array( $admin, 'change_srcset_size_limit' ), 999, 2 );
 		add_filter( 'wp_editor_set_quality', array( $admin, 'change_image_quality' ) );
+		add_filter( 'jpeg_quality', array( $admin, 'change_image_quality' ) );
 		add_action( 'login_head', array( $admin, 'change_login_logo' ) );
 		add_filter( 'login_headerurl', array( $admin, 'change_login_header_url' ) );
 		add_filter( 'login_headertext', array( $admin, 'change_login_header_text' ) );